+++ detector/service/image_preprocessor.py
@@ -5,7 +5,7 @@ import logging
 from detector.model.configuration import ProcessingConfig
 from detector.model.detection_result_code import ResultCode
 from detector.model.exception import DartDetectionError
-from detector.model.image_models import DartImage, DartImagePreprocessed
+from detector.model.image_models import DartImage, DartImagePreprocessed, PreprocessingResult
 from detector.util.file_utils import resize_image
 from detector.yolo.dartboard_cropper import YoloDartBoardImageCropper
 
@@ -32,4 +32,18 @@ class ImagePreprocessor:
             self.logger.info("Cropping model is disabled, skipping image cropping")
 
         resized_image = resize_image(image, self.__config.target_image_size)
-        return DartImagePreprocessed(resized_image, crop_info)
+        return DartImagePreprocessed(resized_image, PreprocessingResult(crop_info=crop_info))
+
+    def preprocess_images_from_preprocessing_result(
+        self, image: DartImage, preprocessing_result: PreprocessingResult
+    ) -> DartImagePreprocessed:
+        """Preprocess images using the provided preprocessing result."""
+        if image is None:
+            raise DartDetectionError(ResultCode.UNKNOWN, details="Input image is None")
+
+        if preprocessing_result.crop_info is not None:
+            self.logger.info("Using provided crop information for preprocessing")
+            image = self.__image_cropper.apply_crop(image, preprocessing_result.crop_info)
+
+        resized_image = resize_image(image, self.__config.target_image_size)
+        return DartImagePreprocessed(resized_image, preprocessing_result)